requirements.txt
# Google APIs dependencies for extract_google_doc_content.py
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
